+++ backend/art-galery/src/main/java/com/unmsm/movil/tecnisis/art_galery/domain/model/ArtisticEvaluation.java
@@ -3,13 +3,15 @@ package com.unmsm.movil.tecnisis.art_galery.domain.model;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.Setter;
+
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 @Builder @Getter @Setter
 public class ArtisticEvaluation {
     private Long id;
     private LocalDate evaluationDate;
-    private Double rating;
+    private BigDecimal rating;
     private String result;
     private Specialist specialist;
     private Request request;
